Refactor: put all opcodes in opcodes.rs
Previously, each opcode had it own file .rs, but now all opcodes
implementations are in opcodes.rs. Also, the opcodes constants that
involved the 4 first bits of instruction moved to the opcodes file too.
