--- src/hardware/opcodes/mod.rs
@@ -1,17 +1,0 @@
-pub mod opcodes_values;
-pub mod utils;
-
-pub mod add;
-pub mod and;
-pub mod br;
-pub mod jmp;
-pub mod jsr;
-pub mod ld;
-pub mod ldi;
-pub mod ldr;
-pub mod lea;
-pub mod not;
-pub mod st;
-pub mod sti;
-pub mod str;
-pub mod trap;
