+++ src/main.rs
@@ -2,8 +2,7 @@ pub mod errors;
 pub mod hardware;
 
 use errors::VmError;
-use hardware::opcodes::opcodes_values;
-use hardware::opcodes::*;
+use hardware::opcodes;
 use hardware::vm::VM;
 use std::{env, fs::File, io::BufReader};
 
@@ -16,47 +15,47 @@ fn execute_instruction(instr: u16, vm: &mut VM) -> Result<(), VmError> {
     let op: u16 = instr >> 12;
 
     match op {
-        opcodes_values::OP_ADD => {
-            add::add(instr, vm)?;
+        hardware::opcodes::OP_ADD => {
+            opcodes::add(instr, vm)?;
         }
-        opcodes_values::OP_AND => {
-            and::and(instr, vm)?;
+        hardware::opcodes::OP_AND => {
+            opcodes::and(instr, vm)?;
         }
-        opcodes_values::OP_NOT => {
-            not::not(instr, vm)?;
+        hardware::opcodes::OP_NOT => {
+            opcodes::not(instr, vm)?;
         }
-        opcodes_values::OP_BR => {
-            br::br(instr, vm)?;
+        hardware::opcodes::OP_BR => {
+            opcodes::br(instr, vm)?;
         }
-        opcodes_values::OP_JMP => {
-            jmp::jmp(instr, vm)?;
+        hardware::opcodes::OP_JMP => {
+            opcodes::jmp(instr, vm)?;
         }
-        opcodes_values::OP_JSR => {
-            jsr::jsr(instr, vm)?;
+        hardware::opcodes::OP_JSR => {
+            opcodes::jsr(instr, vm)?;
         }
-        opcodes_values::OP_LD => {
-            ld::ld(instr, vm)?;
+        hardware::opcodes::OP_LD => {
+            opcodes::ld(instr, vm)?;
         }
-        opcodes_values::OP_LDI => {
-            ldi::ldi(instr, vm)?;
+        hardware::opcodes::OP_LDI => {
+            opcodes::ldi(instr, vm)?;
         }
-        opcodes_values::OP_LDR => {
-            ldr::ldr(instr, vm)?;
+        hardware::opcodes::OP_LDR => {
+            opcodes::ldr(instr, vm)?;
         }
-        opcodes_values::OP_LEA => {
-            lea::lea(instr, vm)?;
+        hardware::opcodes::OP_LEA => {
+            opcodes::lea(instr, vm)?;
         }
-        opcodes_values::OP_ST => {
-            st::st(instr, vm)?;
+        hardware::opcodes::OP_ST => {
+            opcodes::st(instr, vm)?;
         }
-        opcodes_values::OP_STI => {
-            sti::sti(instr, vm)?;
+        hardware::opcodes::OP_STI => {
+            opcodes::sti(instr, vm)?;
         }
-        opcodes_values::OP_STR => {
-            str::str(instr, vm)?;
+        hardware::opcodes::OP_STR => {
+            opcodes::str(instr, vm)?;
         }
-        opcodes_values::OP_TRAP => {
-            trap::trap(instr, vm)?;
+        hardware::opcodes::OP_TRAP => {
+            opcodes::trap(instr, vm)?;
         }
         _ => {} // RTI and RES should not be used
     }
